+++ Package.swift
@@ -266,6 +266,7 @@ let package = Package(
         .target(
             name: "PingFido",
             dependencies: [
+                "PingLogger",
                 "PingCommons",
                 "PingDavinciPlugin",
                 "PingJourneyPlugin"